The ldap url is now a property set in ogp.properties.  The url intercept pattern in security-app-context.xml is changed to provide for a non-https logout and to remove the duplicate login set up by Spring security.
The property-placeholder in WEB-INF/applicationContext.xml was modified
to pick up the properties file in the
classpath*:org/OpenGeoPortal/**/*.properties to allow 'plug-in' jars to
have properties files.

The component scan in appServlet-servlet.xml was changed to a base
package of org.OpenGeoPortal so that any controllers in plug-in jars
would be found.

classpath*:applicationContext.xml was added to the
'contextConfigLocation' parameter in web.xml so that
applicationContext.xml files in plug-in jars would be found by Spring.

bugfix in userInterface.js: fadeIn for main div was missing.