Fixes encoding problems with FGDC download
By default the servlet response writer will use ISO 8859-1, so we
should explicitly set it to UTF-8 in this case. Aside from potential
encoding problems, the content-length header is being misreported in
cases where the UTF-8 encoding differs in byte length than the ISO
8859-1 encoding.

In addition, the content-type should be set to application/xml not
application/octet-stream.